Move consumables timers to their own package
