+++ app/src/main/java/nz/net/speakman/destinyraidtimers/consumables/ConsumablesModule.java
@@ -22,6 +22,8 @@ import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
+import nz.net.speakman.destinyraidtimers.consumables.timers.GlimmerTimer;
+import nz.net.speakman.destinyraidtimers.consumables.timers.TelemetryTimer;
 import nz.net.speakman.destinyraidtimers.consumables.views.GlimmerCountdownView;
 import nz.net.speakman.destinyraidtimers.consumables.views.TelemetryCountdownView;
