+++ app/src/main/java/nz/net/speakman/destinyraidtimers/consumables/views/GlimmerCountdownView.java
@@ -26,9 +26,9 @@ import com.squareup.otto.Subscribe;
 import javax.inject.Inject;
 
 import nz.net.speakman.destinyraidtimers.R;
-import nz.net.speakman.destinyraidtimers.consumables.GlimmerTimer;
-import nz.net.speakman.destinyraidtimers.consumables.GlimmerTimerUpdateEvent;
-import nz.net.speakman.destinyraidtimers.consumables.ConsumablesTimer;
+import nz.net.speakman.destinyraidtimers.consumables.timers.GlimmerTimer;
+import nz.net.speakman.destinyraidtimers.consumables.timers.GlimmerTimerUpdateEvent;
+import nz.net.speakman.destinyraidtimers.consumables.timers.ConsumablesTimer;
 
 /**
  * Created by Adam on 15-03-28.
